move APIs from runtime/js/40_files.js to runtime/js/30_fs.js
